agentcore-runtime-agent/agent_pdz_01/pyproject.toml
[project]
name = "agent-pdz-01"
version = "1.0.0"
description = "Simple AgentCore Runtime agent - PDZ-01"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "httpx",
    "strands-agents",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
